pyproject.toml
[project]
name = "aie-certification-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth-oauthlib>=1.2.2",
    "html2text>=2025.4.15",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
